+++ api/data/network-fees.js
@@ -3,6 +3,9 @@ const fetch = require('node-fetch')
 const Sentry = require('@sentry/node')
 const BigNumber = require('bignumber.js')
 const { fixDecimalsAndRoundUp } = require('../common/numbers.js')
+const database = require('../lib/database')
+const config = require('../config')
+const db = database(config)('')
 
 const transactionTypesSet = new Set([
   'SendTx',
@@ -25,28 +28,30 @@ const EMONEY_GAS_PRICES_ENDPOINT = `http://emoney.validator.network/light/author
 
 let terraTaxRate
 let emoneyGasPrices
+let networkGasPricesDictionary = {}
+let networkGasEstimatesDictionary = {}
 
 const pollForNewFees = async () => {
   const terraTaxRateResponse = await fetch(TERRA_TAX_RATE_ENDPOINT)
-  .then((r) => r.json())
-  .catch((err) => {
-    Sentry.withScope(function (scope) {
-      scope.setExtra('terra tax rate endpoint', TERRA_TAX_RATE_ENDPOINT)
-      Sentry.captureException(err)
+    .then((r) => r.json())
+    .catch((err) => {
+      Sentry.withScope(function (scope) {
+        scope.setExtra('terra tax rate endpoint', TERRA_TAX_RATE_ENDPOINT)
+        Sentry.captureException(err)
+      })
+      return {}
     })
-    return {}
-  })
   const emoneyGasPricesResponse = await fetch(EMONEY_GAS_PRICES_ENDPOINT)
-  .then((r) => r.json())
-  .catch((err) => {
-    Sentry.withScope(function (scope) {
-      scope.setExtra('emoney gas prices endpoint', EMONEY_GAS_PRICES_ENDPOINT)
-      Sentry.captureException(err)
+    .then((r) => r.json())
+    .catch((err) => {
+      Sentry.withScope(function (scope) {
+        scope.setExtra('emoney gas prices endpoint', EMONEY_GAS_PRICES_ENDPOINT)
+        Sentry.captureException(err)
+      })
     })
-  })
   terraTaxRate = Number(terraTaxRateResponse.result)
   emoneyGasPrices = emoneyGasPricesResponse.result.min_gas_prices.map(gasPrice => gasPrice = {
-    denom: gasPrice.denom, 
+    denom: gasPrice.denom,
     price: gasPrice.amount
   })
   setTimeout(async () => {
@@ -54,8 +59,31 @@ const pollForNewFees = async () => {
   }, FEES_POLLING_INTERVAL)
 }
 
+const getNetworkGasPrices = async (networkId) => {
+  const networkGasPrices = await db.getNetworkGasPrices()
+  networkGasPrices.forEach(networkGasPrice => {
+    if (!networkGasPricesDictionary[networkGasPrice.id]) networkGasPricesDictionary[networkGasPrice.id] = []
+    networkGasPricesDictionary[networkGasPrice.id].push({ denom: networkGasPrice.denom,
+      gasPrice: networkGasPrice.gasPrice
+    })
+  })
+  // update emoneyGasPrices
+  networkGasPricesDictionary['emoney-mainnet'] = emoneyGasPrices
+  return networkGasPricesDictionary[networkId] || []
+}
+
+const getNetworkGasEstimates = async (networkId) => {
+  const networkGasEstimates = await db.getNetworkGasEstimates()
+  networkGasEstimates.forEach(networkGasEstimate => {
+    if (!networkGasEstimatesDictionary[networkGasEstimate.id]) networkGasEstimatesDictionary[networkGasEstimate.id] = {}
+    networkGasEstimatesDictionary[networkGasEstimate.id][networkGasEstimate.transactionType] = networkGasEstimate.gasEstimate
+  })
+  return networkGasEstimatesDictionary[networkId]
+}
+
 // run on API launch
 pollForNewFees()
+getNetworkGasEstimates()
 
 const getNetworkTransactionGasEstimates = (networkId, transactionType) => {
   if (transactionType && !transactionTypesSet.has(transactionType)) {
@@ -78,12 +106,6 @@ const getNetworkTransactionGasEstimates = (networkId, transactionType) => {
     : networkGasEstimates.default
 }
 
-const getNetworkGasPrices = (networkId) => {
-  // update emoneyGasPrices
-  networkGasPricesDictionary['emoney-mainnet'] = emoneyGasPrices
-  return networkGasPricesDictionary[networkId]
-}
-
 const getNetworkTransactionChainAppliedFees = (networkId, transactionType) => {
   if (networkId.startsWith('terra') && transactionType === 'SendTx') {
     return {
@@ -93,124 +115,6 @@ const getNetworkTransactionChainAppliedFees = (networkId, transactionType) => {
   } else {
     return null
   }
-}
-
-const terraGasEstimates = {
-    default: 350000,
-    ClaimRewardsTx: 550000
-}
-
-const cosmosGasEstimates = {
-  default: 550000
-}
-
-const emoneyGasEstimates = {
-  default: 300000,
-  SendTx: 75000,
-  StakeTx: 550000,
-  UnstakeTx: 550000,
-  ClaimRewardsTx: 550000,
-  RestakeTx: 550000
-}
-
-const akashGasEstimates = {
-  default: 550000
-}
-
-const polkadotGasEstimates = {
-  default: 0
-}
-
-const kavaGasEstimates = {
-  default: 550000
-}
-
-const networkGasEstimatesDictionary = {
-  'cosmos-hub-mainnet': cosmosGasEstimates,
-  'cosmos-hub-testnet': cosmosGasEstimates,
-  'terra-mainnet': terraGasEstimates,
-  'terra-testnet': terraGasEstimates,
-  'emoney-mainnet': emoneyGasEstimates,
-  'emoney-testnet': emoneyGasEstimates,
-  'akash-testnet': akashGasEstimates,
-  'kusama': polkadotGasEstimates,
-  'polkadot': polkadotGasEstimates,
-  'polkadot-testnet': polkadotGasEstimates,
-  'kava-mainnet': kavaGasEstimates,
-  'kava-testnet': kavaGasEstimates
-}
-
-const cosmosGasPrices = [
-  {
-    denom: 'uatom',
-    price: '0.65e-2'
-  },
-  {
-    denom: 'umuon',
-    price: '0.65e-2'
-  }
-]
-
-const terraGasPrices = [
-  {
-    denom: 'ukrw',
-    price: '0.01'
-  },
-  {
-    denom: 'uluna',
-    price: '0.015'
-  },
-  {
-    denom: 'umnt',
-    price: '0.01'
-  },
-  {
-    denom: 'usdr',
-    price: '0.01'
-  },
-  {
-    denom: 'uusd',
-    price: '0.01'
-  }
-]
-
-const kavaGasPrices = [
-  {
-    denom: 'ukava',
-    price: '0.1'
-  },
-  {
-    denom: 'usdx',
-    price: '0.1'
-  },
-  {
-    denom: 'bnb',
-    price: '0.1'
-  },
-]
-
-const akashGasPrices = [
-  {
-    denom: 'akash',
-    price: '0.01'
-  }
-]
-
-const polkadotGasPrices = []
-
-let networkGasPricesDictionary = {
-  'cosmos-hub-mainnet': cosmosGasPrices,
-  'cosmos-hub-testnet': cosmosGasPrices,
-  'terra-mainnet': terraGasPrices,
-  'terra-testnet': terraGasPrices,
-  'emoney-mainnet': emoneyGasPrices,
-  'emoney-testnet': emoneyGasPrices,
-  'kava-mainnet': kavaGasPrices,
-  'kava-testnet': kavaGasPrices,
-  'akash-testnet': akashGasPrices,
-  'kusama': polkadotGasPrices,
-  'polkadot': polkadotGasPrices,
-  'polkadot-testnet': polkadotGasPrices,
 }
 
 const getPolkadotMessage = async (messageType, senderAddress, message, network, networkSource) => {
@@ -243,12 +147,12 @@ const getPolkadotFee = async ({ messageType, message, senderAddress, network, ne
     if (message.amounts) {
       const { amounts } = message
       amount = amounts[0]
-    }  
+    }
     return {
       denom: (amount && amount.denom) || network.stakingDenom,
       amount: viewFees
     }
-  } catch(error) {
+  } catch (error) {
     Sentry.captureException(error)
     // back up plan. Send most common fee
     // TODO: check it this is the same for Polkadot network
@@ -295,7 +199,7 @@ const getTransactionAmount = (message, feeDenom) => {
   }
   // check if there is an amounts field
   if (message.amounts) {
-    return message.amounts.find(({denom}) => denom === feeDenom).amount
+    return message.amounts.find(({ denom }) => denom === feeDenom).amount
   }
   return 0
 }
@@ -333,14 +237,15 @@ const selectAlternativeFee = (balances, feeDenom, gasEstimate) => {
 
 const getCosmosFee = async (network, cosmosSource, senderAddress, messageType, message, gasEstimate) => {
   // query for this address balances
-  const balances = await cosmosSource.getBalancesFromAddress(senderAddress, '', network)
+  const [balances, accountInfo] = await Promise.all([
+    cosmosSource.getBalancesFromAddress(senderAddress, '', network),
+    cosmosSource.getAccountInfo(senderAddress)
+  ])
   const feeDenom = getFeeDenomFromMessage(message, network)
-  const gasPrice = BigNumber(
-    getNetworkGasPrices(network.id).find(({ denom }) => {
-      const coinLookup = network.getCoinLookup(network, denom)
-      return coinLookup ? coinLookup.viewDenom === feeDenom : false
-    }).price
-  )
+  const gasPrice = BigNumber(networkGasPricesDictionary[network.id].find(({ denom }) => {
+    const coinLookup = network.getCoinLookup(network, denom)
+    return coinLookup ? coinLookup.viewDenom === feeDenom : false
+  }).gasPrice)
     .times(
       network.getCoinLookup(network, feeDenom, `viewDenom`)
         .chainToViewConversionFactor
@@ -354,12 +259,16 @@ const getCosmosFee = async (network, cosmosSource, senderAddress, messageType, m
   let estimatedFee = {
     amount: String(
       chainAppliedFees && chainAppliedFees.rate > 0
-        ? fixDecimalsAndRoundUp(BigNumber(transactionAmount).times(chainAppliedFees.rate).toNumber(), 6)
-        : fixDecimalsAndRoundUp(gasEstimate * gasPrice, 6)
-    ),
+        ? fixDecimalsAndRoundUp(BigNumber(transactionAmount).times(chainAppliedFees.rate).plus(BigNumber(gasEstimate * gasPrice)).toNumber(), 6)
+        : fixDecimalsAndRoundUp(gasEstimate * gasPrice, 6)    
+      ),
     denom: feeDenom
   }
-  const selectedBalance = balances.find(({denom}) => denom === feeDenom) || { amount: 0, denom: feeDenom }
+  const selectedBalance = balances.find(({ denom }) => denom === feeDenom) || { amount: 0, denom: feeDenom }
+  // HACK, should check the not vested balance for fees
+  if (accountInfo.vestingAccount) {
+    selectedBalance.amount = 0
+  }
   if (
     Number(transactionAmount) + Number(estimatedFee.amount) >
     Number(selectedBalance.amount) &&
@@ -384,12 +293,13 @@ const getCosmosFee = async (network, cosmosSource, senderAddress, messageType, m
   }
 }
 
-module.exports = { 
-  getNetworkTransactionGasEstimates, 
-  getNetworkTransactionChainAppliedFees, 
-  getNetworkGasPrices, 
-  getPolkadotFee, 
-  getPolkadotMessage, 
-  getFeeDenomFromMessage, 
+module.exports = {
+  getNetworkTransactionGasEstimates,
+  getNetworkTransactionChainAppliedFees,
+  getNetworkGasPrices,
+  getNetworkGasEstimates,
+  getPolkadotFee,
+  getPolkadotMessage,
+  getFeeDenomFromMessage,
   getCosmosFee,
 }