+++ api/lib/reducers/cosmosV2-reducers.js
@@ -55,7 +55,7 @@ function sendDetailsReducer(message, reducers, network) {
   return {
     from: [message.from_address],
     to: [message.to_address],
-    amount: reducers.coinReducer(message.amount[0], coinLookup)
+    amount: reducers.coinReducer(message.amount[0], coinLookup, network)
   }
 }
 
@@ -63,7 +63,7 @@ function stakeDetailsReducer(message, reducers, network) {
   const coinLookup = network.getCoinLookup(network, message.amount.denom)
   return {
     to: [message.validator_address],
-    amount: reducers.coinReducer(message.amount, coinLookup)
+    amount: reducers.coinReducer(message.amount, coinLookup, network)
   }
 }
 
@@ -72,7 +72,7 @@ function restakeDetailsReducer(message, reducers, network) {
   return {
     from: [message.validator_src_address],
     to: [message.validator_dst_address],
-    amount: reducers.coinReducer(message.amount, coinLookup)
+    amount: reducers.coinReducer(message.amount, coinLookup, network)
   }
 }
 
@@ -80,7 +80,7 @@ function unstakeDetailsReducer(message, reducers, network) {
   const coinLookup = network.getCoinLookup(network, message.amount.denom)
   return {
     from: [message.validator_address],
-    amount: reducers.coinReducer(message.amount, coinLookup)
+    amount: reducers.coinReducer(message.amount, coinLookup, network)
   }
 }
 
@@ -142,7 +142,11 @@ function submitProposalDetailsReducer(message, reducers, network) {
     proposalType: message.content.type,
     proposalTitle: message.content.value.title,
     proposalDescription: message.content.value.description,
-    initialDeposit: reducers.coinReducer(message.initial_deposit[0], coinLookup)
+    initialDeposit: reducers.coinReducer(
+      message.initial_deposit[0],
+      coinLookup,
+      network
+    )
   }
 }
 
@@ -157,7 +161,7 @@ function depositDetailsReducer(message, reducers, network) {
   const coinLookup = network.getCoinLookup(network, message.amount[0].denom)
   return {
     proposalId: message.proposal_id,
-    amount: reducers.coinReducer(message.amount[0], coinLookup)
+    amount: reducers.coinReducer(message.amount[0], coinLookup, network)
   }
 }
 
@@ -272,14 +276,14 @@ function transactionReducerV2(network, transaction, reducers) {
     ) {
       fees = transaction.tx.value.fee.amount.map((coin) => {
         const coinLookup = network.getCoinLookup(network, coin.denom)
-        return coinReducer(coin, coinLookup)
+        return coinReducer(coin, coinLookup, network)
       })
     } else {
       const coinLookup = network.getCoinLookup(
         network,
         transaction.tx.value.fee.amount.denom
       )
-      fees = [coinReducer(transaction.tx.value.fee.amount, coinLookup)]
+      fees = [coinReducer(transaction.tx.value.fee.amount, coinLookup, network)]
     }
     // We do display only the transactions we support in Lunie
     const filteredMessages = transaction.tx.value.msg.filter(
@@ -333,7 +337,14 @@ function transactionReducerV2(network, transaction, reducers) {
             ? transaction.logs[index].log || transaction.logs[0] // failing txs show the first logs
             : transaction.logs[0].log || ''
           : JSON.parse(JSON.stringify(transaction.raw_log)).message,
-      involvedAddresses: uniq(reducers.extractInvolvedAddresses(transaction))
+      involvedAddresses: Array.isArray(transaction.logs)
+        ? uniq(
+            reducers.extractInvolvedAddresses(
+              transaction.logs.find(({ msg_index }) => msg_index === index)
+                .events
+            )
+          )
+        : []
     }))
     return returnedMessages
   } catch (error) {
@@ -353,7 +364,7 @@ function transactionsReducerV2(network, txs, reducers) {
   // here we filter out all transactions related to validators
   return reversedTxs.reduce((collection, transaction) => {
     return collection.concat(
-      transactionReducerV2(network, transaction, reducers)
+      reducers.transactionReducerV2(network, transaction, reducers)
     )
   }, [])
 }
@@ -393,13 +404,17 @@ function validatorReducer(networkId, signedBlocksWindow, validator) {
       ? validator.signing_info.start_height
       : undefined,
     uptimePercentage:
-      1 -
-      Number(
-        validator.signing_info
-          ? validator.signing_info.missed_blocks_counter
-          : 0
-      ) /
-        Number(signedBlocksWindow),
+      validator.signing_info &&
+      validator.signing_info.missed_blocks_counter &&
+      signedBlocksWindow
+        ? 1 -
+          Number(
+            validator.signing_info
+              ? validator.signing_info.missed_blocks_counter
+              : 0
+          ) /
+            Number(signedBlocksWindow)
+        : undefined,
     tokens: atoms(validator.tokens),
     commissionUpdateTime: validator.commission.update_time,
     commission: Number(validator.commission.commission_rates.rate).toFixed(6),
@@ -412,31 +427,25 @@ function validatorReducer(networkId, signedBlocksWindow, validator) {
   }
 }
 
-function extractInvolvedAddresses(transaction) {
-  // If the transaction has failed, it doesn't get tagged
-  if (!Array.isArray(transaction.events)) return []
-
+function extractInvolvedAddresses(events) {
   // extract all addresses from events that are either sender or recipient
-  const involvedAddresses = transaction.events.reduce(
-    (involvedAddresses, event) => {
-      const senderAttributes = event.attributes
-        .filter(({ key }) => key === 'sender')
-        .map((sender) => sender.value)
-      if (senderAttributes.length) {
-        involvedAddresses = [...involvedAddresses, ...senderAttributes]
-      }
+  const involvedAddresses = events.reduce((involvedAddresses, event) => {
+    const senderAttributes = event.attributes
+      .filter(({ key }) => key === 'sender')
+      .map((sender) => sender.value)
+    if (senderAttributes.length) {
+      involvedAddresses = [...involvedAddresses, ...senderAttributes]
+    }
 
-      const recipientAttribute = event.attributes.find(
-        ({ key }) => key === 'recipient'
-      )
-      if (recipientAttribute) {
-        involvedAddresses.push(recipientAttribute.value)
-      }
+    const recipientAttribute = event.attributes.find(
+      ({ key }) => key === 'recipient'
+    )
+    if (recipientAttribute) {
+      involvedAddresses.push(recipientAttribute.value)
+    }
 
-      return involvedAddresses
-    },
-    []
-  )
+    return involvedAddresses
+  }, [])
   return involvedAddresses
 }
 
@@ -461,6 +470,7 @@ module.exports = {
   // CosmosV0 Reducers
   ...cosmosV0Reducers,
   transactionsReducerV2,
+  transactionDetailsReducer,
   transactionReducerV2,
   sendDetailsReducer,
   stakeDetailsReducer,