+++ app/src/components/common/TmBalance.vue
@@ -73,8 +73,9 @@
                 value=""
                 :selected="preferredCurrency"
                 hidden
-                >{{ preferredCurrency }}</option
               >
+                {{ preferredCurrency }}
+              </option>
               <option value="USD">USD</option>
               <option value="EUR">EUR</option>
               <option value="CAD">CAD</option>
@@ -555,6 +556,10 @@ select option {
 
   .currency-selector {
     display: none;
+  }
+
+  .reward-bar {
+    margin: 2rem 2rem;
   }
 }