+++ app/src/components/governance/PageProposals.vue
@@ -55,16 +55,14 @@
     <ModalPropose
       v-if="parameters && Object.keys(parameters).length > 0"
       ref="modalPropose"
-      :denom="parameters.depositDenom"
+      :denom="parameters.depositDenom || currentNetwork.stakingDenom"
       @success="() => afterPropose()"
     />
 
     <div v-if="!$apollo.loading && proposals.length === 0">
       <div>
         <TmDataMsg icon="gavel">
-          <div slot="title">
-            No Governance Proposals
-          </div>
+          <div slot="title">No Governance Proposals</div>
           <div slot="subtitle">
             There are currently no governance proposals to display. Click the
             'Create Proposal' button to submit the first network proposal!